chore: 로그아웃 부분 제거로 스토리북 CI 깨지는것 방지
