+++ src/components/layout/Header/index.tsx
@@ -1,34 +1,19 @@
 'use client';
 
-import { useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
 import { userQueries } from '@/api/queries/userQueries';
 import LogoDefault from '@/assets/icons/logo-default.svg?react';
 import LogoLarge from '@/assets/icons/logo-large.svg?react';
 import Dropdown from '@/components/ui/Dropdown';
 import UserAvatar from '@/components/ui/UserAvatar';
 
+// import { useSignout } from './useSignout';
+
 export default function Header() {
-  const queryClient = useQueryClient();
-  const router = useRouter();
   const { data: user, isLoading } = useQuery(userQueries.me());
   const isLoggedIn = !isLoading && !!user;
-
-  const handleSignout = async () => {
-    try {
-      const response = await fetch('/api/auth/signout', {
-        method: 'POST',
-      });
-
-      if (response.ok) {
-        queryClient.invalidateQueries({ queryKey: userQueries.me().queryKey });
-        router.push('/signin');
-      }
-    } catch (error) {
-      console.error('Signout failed:', error);
-    }
-  };
+  // const handleSignout = useSignout();
 
   return (
     <header className="tablet:h-15 tablet:px-6 sticky top-0 z-50 h-14 w-full border-b border-b-gray-600 bg-gray-800 px-4">
@@ -61,9 +46,7 @@ export default function Header() {
                   <Dropdown.Item>
                     <Link href="/my">마이페이지</Link>
                   </Dropdown.Item>
-                  <Dropdown.Item onClick={handleSignout}>
-                    로그아웃
-                  </Dropdown.Item>
+                  <Dropdown.Item>로그아웃</Dropdown.Item>
                 </Dropdown.Content>
               </Dropdown>
             </>
